Initial Travis CI configuration
At least to start, I want Travis CI to run autogen to create the
`configure` script and `Makefile.in` files, run `./configure && make
check`, and commit and push back any `autogen` products back to the
repository.  This way, `configure` and `Makefile.in`s are always the
result of a working build, up to date, and generated from a know version
of autoconf and automake, not just what is installed on the last dev's
machine.