Add initial fontmake autoconfiguration
This is a first take at exposing configure parameters for `python`,
`pip`, and `fontmake` as environment variables `PYTHON`, `PIP`, and
`FONTMAKE`, similar to how paths to `gcc` and `g++` can be set as
environment variables `CC` and `CXX`.  When `FONTMAKE=pip`, `make` is
instructed to install `fontmake` using `pip` to a build-local `deps`
directory.