Add new configure arg: --with-font-srcdir
Configure will now search `..`, `../src`, and `../source`, in that
order, for `*.ufo` files.  If found, that directory is set as the
`font_srcdir`, which can be referred to in make.  That should cover most
cases when autofont is used as a subdirectory in a font project.  If the
source is not in one of those directories, it can be specified with the
`--with-font-srcdir` configure argument.