Not finding a font source dir is not a failure
This is kind of a subtle requirement of autofont.  Configure should
always be able to run so that releases can be made with `make dist`.  So
this commit makes it so not finding any font sources is just a warning,
and missing dependencies are only a failure if font sources have been
found.