Separate concepts of font_srcdir and font_ufodir
The `font_ufodir` must exist under `font_srcdir`, and that is (mostly)
enforced by not being able to configure anything but a directory name of
the `ufodir`, like `src` or `source` or `ufos` or `.`.  By default,
configure will check directory names `.`, `src`, and `source` under
`font_srcdir` for UFOs.  It otherwise behaves the same as d82fd09.

The reason for this separation is because:

1. I want to be able to make font source releases, and for that make
will need to know the root directory of the font project.

2. make still needs to know the directory containing the UFO sources so
it can figure out what to make, and establish dependencies between the
sources and the compiled targets.

Mostly, no one will ever have to interact with either the
`--with-font-srcdir` or `--with-font-ufodirname` arguments.  And they
can be made configurable in the font project so projects that have a
weird directory structure can configure autofont and then still no one
will have to use these configure arguements.