Only let configure run from other directories
This is maybe a controvesial-sounding change, because in most autotools
projects, you do `cd project; ./configure`, and this would prevent that.
That usage pattern doesn't really apply to the way I'm intending
autofont to be consumed, which is for it to be in a subdirectory anyway.
Thus the typical usage pattern will be `cd project; autofont/configure`.
This change enforces that.

As a nice side effect, this guarantees that VPATH builds will always
work (which running `configure` in the same directory prevents), so unit
can work, and it guarantees the `GNUmakefile` in the root of the
autofont project is free for autofont developer's rules.