Refactor pip detection
In preparation for using the Python ufoLib to do some configuration
checks, change the way pip is requested/detected so that there is a
single argument, `--enable-build-dependencies`, that can be used to turn
on installation of missing dependencies.  This also removes the PIP
environment variable in favor of running `$PYTHON -m pip`, which will
prevent a situation where incompatible pip and python versions are
specified.