Check for and possibly install ufoLib
I intend to use the ufoLib Python module to read a UFO file at
configure-time to pull family name and maybe other metadata information
out and make configure-time decisions based on it.

As an example, I want to be able to specify the Automake TTF install-dir
on Linux as $prefix/fonts/$font_family_name, but that is most clearly
asserted at configure-time.  Most projects would know their asset names
so this is not a problem; but for Autofont, which should work as a drop
in to any project without configuration, we need to detect these kind of
names.