Detect source typeface family name
Normally build tools know the name of the project they're building.
Autofont has the additional task of figuring out what it's building.  We
can't trust the names of the ufo files because they can be arbitrarily
named.  We have to pull the names out of the UFO metadata.  To make
matters more complicated, a project may have many different typeface
names (think "Deja Vu Sans", "Deja Vu Serif", "Deja Vu Sans Mono", etc).

This commit delivers a Python script that reads the font family name out
of all a project's UFO files, and prints the common name between them.
In the Deja Vu example, it outputs "Deja Vu".  Autofont will use that as
the basis for installation directory names and font renaming support
when we get to it.