configure: Set installdir based on OS/--with-installdir arg
Add a new configure argument, `--with-installdir=DIR`, to set where to
install the fonts with `make install`, or `--without-installdir` to
disable installation.  If unspecified, try to set the installdir based
on the host OS (which is autoconf speak for the target OS, which default
s to the build host OS, but can be overridden with the `--host`
argument).  For example, on Linux, this is
`$prefix/share/fonts/$family_name`.  MacOS is something like
`/Library/Fonts`, but I need to test on an actual macOS system to see
how that behaves before I add support for it.