Add configure support for ttfautohint
A new argument is added to `configure` called `--with-hints` to let the
user control what hinting strategy to apply to the resulting TTFs.
`configure` will attempt to find `ttfautohint` in the build environment,
and allow the path to the executable to be set with the `TTFAUTOHINT`
environment variable, or left to compile and install at build time with
the `--enable-build-environment` option.

Accompanying unit tests and documentation are also provided.

What this does not yet do is add ttfautohint support to the makefile.